refactor: split admin management sections
